Supabase DB connection successful
Successfully connected to the Supabase PostgreSQL database using the supabase_flutter package. Called a stored procedure with parameters to show successful connection in a snackbar on the home_screen

Added the dotenv package to the project and created the .env file to store the supabase credentials. The .env file was not added to the .gitignore, so it will be stored in the repo for use across the project.